combined both progress servlets into ProgressServlet.java
