--- capstone/fitness/src/main/java/com/google/sps/servlets/GetProgressServlet.java
@@ -1,37 +1,0 @@
-package com.google.sps.servlets;
-
-import com.google.appengine.api.datastore.DatastoreService;
-import com.google.appengine.api.datastore.DatastoreServiceFactory;
-import com.google.appengine.api.datastore.Entity;
-import com.google.appengine.api.datastore.PreparedQuery;
-import com.google.appengine.api.datastore.Query;
-import com.google.appengine.api.users.UserService;
-import com.google.appengine.api.users.UserServiceFactory;
-import com.google.sps.util.DataHelper;
-import java.io.IOException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-/** Servlet that returns some example content. TODO: modify this file to handle comments data */
-@WebServlet("/progress")
-public class GetProgressServlet extends HttpServlet {
-
-  @Override
-  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    
-    // Get the users email
-    UserService userService = UserServiceFactory.getUserService();
-    String userEmail = userService.getCurrentUser().getEmail();
- 
-    Entity user=DataHelper.getUser();
-
-    // Get the users progress from datastore
-    String progressString = (String) user.getProperty(DataHelper.PROGRESS_PROPERTY);
-
-    //Give the output in JSON format
-    response.setContentType("application/json");
-    response.getWriter().println(progressString);
-  }
-}
